Adding the temporary user management:   - Refactored the hierarchy of views for having a base class with     create and update datetimes.   - Discovered that the naive usage of session to register votes and     other actions does not work as the admin-site erases the session on     login.   - Came up with a middle ware to handle the creation of a temporary     user (TemporaryUser model).   - Refactored the views to have them all in the class format. The     syntax is a bit heavier but eventually it will help having     consistency among them.
Refactored the temporary user into a visitor model. Having a visitor identifier (Visitor.pk field) will help to register the un-logged user actions but also keep statistics in the future. It is a tradeoff design between the Session (that is too limited and interlly used by Django) and the TemporaryUser (which name is misleading compared to the registered user).
 - As a consequence, the migrations needed to be refactored. The demoData.json also needed to be amended.

Adding the midleware for the visitor id to be registered in the session.